symbolfield-crew/main.py
from crewai import Agent, Task, Crew, Process
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv() # Load env vars from .env in root

# Mock config loader for v0.5 scaffold
def load_config(path):
    # In real impl, read yaml
    return {}

class SymbolFieldCrew:
    def __init__(self):
        self.agents_config = load_config('config/agents.yaml')
        self.tasks_config = load_config('config/tasks.yaml')

    def context_manager(self):
        return Agent(
            role='Context Manager',
            goal='Sync context spine',
            backstory='Librarian of SymbolField',
            verbose=True,
            allow_delegation=False,
            # tools=[SupabaseTool(), FileSystemTool()]
        )

    def qa_guardian(self):
        return Agent(
            role='QA Guardian',
            goal='Prevent regressions',
            backstory='The Shield',
            verbose=True,
            # tools=[PlaywrightTool(), StaticAnalysisTool()]
        )

    def run(self):
        # Define Agents
        manager = self.context_manager()
        guardian = self.qa_guardian()

        # Define Tasks
        sync_task = Task(
            description='Sync docs to Supabase',
            agent=manager,
            expected_output='Log of synced files'
        )
        
        audit_task = Task(
            description='Run Smoke/Regression Tests',
            agent=guardian,
            expected_output='Audit Report MD'
        )

        crew = Crew(
            agents=[manager, guardian],
            tasks=[sync_task, audit_task],
            verbose=2,
            process=Process.sequential
        )

        result = crew.kickoff()
        return result

if __name__ == "__main__":
    print("Starting SymbolField Crew (Audit Loop)...")
    # symbol_field_crew = SymbolFieldCrew()
    # symbol_field_crew.run()
    print("Crew scaffold ready. Install requirements to run.")
