symbolfield-crew/requirements.txt
crewai
langchain
playwright
supabase
pyyaml
python-dotenv
